requirements.txt
discord.py[voice]
wavelink
validators
git+https://github.com/Rapptz/discord-ext-menus